+++ engine/src/main/java/me/molybdenum/ambience_mini/engine/client/music/MusicPlayer.java
@@ -7,8 +7,11 @@ import me.molybdenum.ambience_mini.engine.shared.utils.Pair;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.*;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 import java.util.Stack;
+import java.util.function.Consumer;
 
 public class MusicPlayer {
     private final Stack<Pair<Integer, Channel>> stack = new Stack<>(); // Priorities and channels
@@ -17,6 +20,8 @@ public class MusicPlayer {
     private final Runnable onRanToEnd;
 
     private float currentMusicVolume;
+
+    private final List<Consumer<NowPlaying>> nowPlayingListeners = new ArrayList<>();
 
 
     public MusicPlayer(MusicProvider musicProvider, Runnable onRanToEnd) {
@@ -70,10 +75,9 @@ public class MusicPlayer {
         synchronized (stack) {
             try {
                 stopAllAbove(priority-1, doFade);
-                Channel channel = createChannel(music);
+                Channel channel = createChannel(music, currentMusicVolume);
                 stack.push(new Pair<>(priority, channel));
-                channel.setVolume(currentMusicVolume);
-                channel.resume(doFade);
+                innerResume(channel, doFade);
             } catch (FileNotFoundException e) {
                 throw new RuntimeException("File '" + music.path() +  "' not found. Fix your Ambience config!", e);
             }
@@ -91,12 +95,11 @@ public class MusicPlayer {
         }
     }
 
-    private Channel createChannel(Music music) throws FileNotFoundException {
+    private Channel createChannel(Music music, float volume) throws FileNotFoundException {
         try {
-            return new Channel(
-                    music,
-                    AmDecoder.of(new MusicInstance(musicProvider, music))
-            );
+            Channel channel = new Channel(music, AmDecoder.of(new MusicInstance(musicProvider, music)));
+            channel.setVolume(volume);
+            return channel;
         }
         catch (Exception ex) {
             throw new RuntimeException("Could not create audio channel for music-file '" + music.path() + "'", ex);
@@ -121,10 +124,21 @@ public class MusicPlayer {
     public void resume(boolean doFade) {
         synchronized (stack) {
             if (!stack.empty())
-                stack.peek().right().resume(doFade);
+                innerResume(stack.peek().right(), doFade);
         }
     }
 
+    private void innerResume(Channel channel, boolean doFade) {
+        notifyListeners(channel);
+        channel.resume(doFade);
+    }
+
+    public void stopAll() {
+        synchronized (stack) {
+            stack.forEach(pair -> pair.right().stopAndClose(false));
+            stack.clear();
+        }
+    }
 
     public boolean isPlaying() {
         synchronized (stack) {
@@ -133,10 +147,24 @@ public class MusicPlayer {
     }
 
 
-    public void stopAll() {
-        synchronized (stack) {
-            stack.forEach(pair -> pair.right().stopAndClose(false));
-            stack.clear();
-        }
+    private void notifyListeners(Channel channel) {
+        var title = channel.getMusicTitle();
+        var titleOrPath = title == null ? channel.getMusicPath() : title;
+        var author = channel.getMusicAuthor();
+        var nowPlaying = new NowPlaying(titleOrPath, author);
+
+        for (var listener : nowPlayingListeners)
+            listener.accept(nowPlaying);
     }
+
+    public void addListener(Consumer<NowPlaying> listener) {
+        nowPlayingListeners.add(listener);
+    }
+
+    public void removeListener(Consumer<NowPlaying> listener) {
+        nowPlayingListeners.remove(listener);
+    }
+
+
+    public record NowPlaying(String titleOrPath, String author) { }
 }